requirements.txt
eventlet
Flask-SocketIO

#codetiming
pytest
pytest-cov
flake8
black
spotipy 
